Move RoomManagerModule to sdk package
