+++ app/src/internal/java/com/twilio/video/app/VideoApplicationComponent.java
@@ -19,10 +19,10 @@ package com.twilio.video.app;
 import com.twilio.video.app.auth.AuthModule;
 import com.twilio.video.app.data.DataModule;
 import com.twilio.video.app.data.api.VideoAppServiceModule;
+import com.twilio.video.app.sdk.RoomManagerModule;
 import com.twilio.video.app.ui.ScreenSelectorModule;
 import com.twilio.video.app.ui.login.LoginActivityModule;
 import com.twilio.video.app.ui.room.RoomActivityModule;
-import com.twilio.video.app.ui.room.RoomManagerModule;
 import com.twilio.video.app.ui.room.VideoServiceModule;
 import com.twilio.video.app.ui.settings.SettingsActivityModule;
 import com.twilio.video.app.ui.settings.SettingsFragmentModule;
